Modify: footer 컴포넌트 브랜치명 수정 및 상수 데이터 분리
