+++ src/components/Footer/mockData.js
@@ -1,0 +1,90 @@
+const UTILITIES = [
+  {
+    id: 1,
+    util: '이용약관',
+  },
+  {
+    id: 2,
+    util: '개인정보처리방침',
+  },
+  {
+    id: 3,
+    util: '구매안전서비스 가입사실 확인',
+  },
+];
+
+const WORKING_TIME = [
+  {
+    id: 1,
+    timeDivision: '평일',
+    Specifictime: '09:00 ~ 18:00',
+  },
+  {
+    id: 2,
+    timeDivision: '점심',
+    Specifictime: '09:00 ~ 18:00',
+  },
+  {
+    id: 3,
+    timeDivision: '휴무',
+    Specifictime: '토/일/공휴일은 휴무',
+  },
+];
+
+const COMPANY_INFO = [
+  {
+    id: 1,
+    title: '상호',
+    value: '(주)학산',
+  },
+  {
+    id: 2,
+    title: '대표이사',
+    value: '김용덕',
+  },
+  {
+    id: 3,
+    title: '주소',
+    value: '서울시 강남구 테헤란로 427',
+  },
+  {
+    id: 4,
+    title: '통신판번호',
+    value: '2003-강원강릉-0085',
+  },
+  {
+    id: 5,
+    title: '계좌번호',
+    value: '신한은행 110-474-658716 (주)학산',
+  },
+  {
+    id: 6,
+    title: '사업자등록번호',
+    value: '012-34-56789',
+  },
+  {
+    id: 7,
+    title: '개인정보보호책임자',
+    value: '김용덕',
+  },
+  {
+    id: 8,
+    title: '원두도매문의',
+    value: '010-8765-4321',
+  },
+  {
+    id: 9,
+    title: '매장이용문의',
+    value: '02-1234-5678',
+  },
+  {
+    id: 10,
+    title: '레스토랑문의',
+    value: '010-1234-5678',
+  },
+  {
+    id: 11,
+    title: '이메일',
+    value: 'info@kleenex.wecode.io',
+  },
+];
